+++ packages/tools/src/replaceImageUrls.ts
@@ -4,6 +4,7 @@
 import qq from '@hugsmidjan/qj/qq';
 
 import { asDiv } from './_cleanup/serverDOM';
+import { FILE_SERVER } from './constants';
 import { HTMLText, RegName } from './types';
 
 const { FILE_UPLOAD_KEY_PUBLISH, REGULATIONS_API_URL } = process.env;
@@ -23,8 +24,21 @@ export const replaceImageUrls = async (html: HTMLText, regName: RegName) => {
     const urls: Array<string> = [];
 
     qq('img', root).forEach((img) => {
+      // Collect all image src URLs as those are an intrinsic part of
+      // the regulation's content and need to be "burned into" the
+      // regulation as hard as possible, to avoid images rotting/breaking
       const src = img.getAttribute('src');
       src && urls.push(src);
+    });
+
+    qq('a', root).forEach((link) => {
+      const href = link.getAttribute('href');
+      // Collect link hrefs that point to the file-server,
+      // as those might have been uploaded during regulation draft authoring.
+      // Those resources need to be moved to a public path on the file-server
+      if (href && FILE_SERVER.endsWith('//' + link.host)) {
+        urls.push(href);
+      }
     });
 
     const sourcesMap: SourcesMap = await fetch(